some followers endpoints and GhActions

+++ src/main/java/austral/ingsis/jjuserservice/service/UserService.java
@@ -1,9 +1,12 @@
 package austral.ingsis.jjuserservice.service;
 
+import austral.ingsis.jjuserservice.dto.UpdateUserDto;
 import austral.ingsis.jjuserservice.dto.UserDto;
+import austral.ingsis.jjuserservice.exception.UserNotFoundException;
 import austral.ingsis.jjuserservice.model.UserDao;
 import austral.ingsis.jjuserservice.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
@@ -11,6 +14,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -18,7 +22,7 @@ public class UserService {
 
     @PersistenceContext
     private EntityManager entityManager; //might not be needed
-    private UserRepository userRepository;
+    private final UserRepository userRepository;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
     @Autowired
@@ -40,11 +44,41 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
+        Optional<UserDao> user = this.userRepository.findById(id);
+        if(user.isEmpty()) {
+            throw new UserNotFoundException("User of id: " + id  + "not found.");
+        }
         this.userRepository.deleteById(id);
     }
 
-    //update
+    public UserDto updateUser(UpdateUserDto dto) {
+        Optional<UserDao> user = this.userRepository.findById(dto.getId());
 
-    //login
+        if(user.isEmpty()) {
+          throw new UserNotFoundException("User of id: " + dto.getId() + "not found.");
+        }
 
+        UserDao toSave = user.get();
+
+        toSave.setEmail(dto.getEmail());
+        toSave.setFirstName(dto.getFirstName());
+        toSave.setLastName(dto.getLastName());
+
+        if(dto.getPassword() != null && !dto.getPassword().isEmpty()) {
+            toSave.setPassword(bCryptPasswordEncoder.encode(dto.getPassword()));
+        }
+
+        return this.userRepository.save(toSave).toUserDto();
+    }
+
+    public UserDto getUserById(Long id) {
+
+        Optional<UserDao> optUser = this.userRepository.findById(id);
+
+       if(optUser.isEmpty()) {
+           throw new UserNotFoundException("User of id: " + id.toString() + "not found.");
+       }
+
+       return optUser.get().toUserDto();
+    }
 }
